Retrain inception with more images

+++ machine_learning_recipes/png2jpg.py
@@ -4,12 +4,17 @@ from os import listdir
 import string
 import random
 
-file_path = '/Users/jo/Desktop/data_science/support_vector_machine/machine_learning_recipes/tf_files/flower_photos/2011_hyundai_accent_gl'
-
+year = 2015
+file_path = f'/Users/jo/Desktop/data_science/support_vector_machine/machine_learning_recipes/tf_files/flower_photos/{year}_hyundai_accent_gls'
+count = 0
 for path in listdir(file_path):
-    im = Image.open(f'{file_path}/{path}')
-    rgb_im = im.convert('RGB')
-    random_name = ''.join(random.choices(
-        string.ascii_uppercase + string.digits, k=20))
-    rgb_im.save(f'{file_path}/{random_name}.jpg')
-    os.remove(f'{file_path}/{path}')
+    if str(year) in path:
+        continue
+    if '.png' in path or '.jpg' in path:
+        im = Image.open(f'{file_path}/{path}')
+        rgb_im = im.convert('RGB')
+        random_name = ''.join(random.choices(
+            string.ascii_uppercase + string.digits, k=20))
+        rgb_im.save(f'{file_path}/{year}_{random_name}.jpg')
+        os.remove(f'{file_path}/{path}')
+# print(count)
